refactor: 불필요한 플레이어 추상화 제거
