+++ src/main/java/blackjack/domain/card/Cards.java
@@ -4,7 +4,6 @@ import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 public class Cards {
     private static final int MAX_SCORE = 21;
