+++ src/main/java/blackjack/domain/gamer/Dealer.java
@@ -2,7 +2,7 @@ package blackjack.domain.gamer;
 
 import blackjack.domain.card.Card;
 
-public class Dealer extends Gamer {
+public class Dealer extends Player {
     public static final int HIT_UPPER_BOUND = 16;
     private static final String name = "딜러";
